requirements.txt
aiogram==3.20.0.post0
aiosqlite==0.20.0
httpx==0.28.1
python-dotenv==1.0.1
